requirements.txt
google-cloud-bigquery>=3.11.0
google-genai>=0.2.0
google-api-python-client>=2.95.0
google-auth>=2.22.0
google-cloud-storage>=2.10.0
